split plugin module
